:truck: move `internal/pipelines` to `internal/data/pipelines`
