+++ api/ingest_log.go
@@ -9,7 +9,7 @@ import (
 
 	"link-society.com/flowg/internal/data/auth"
 	"link-society.com/flowg/internal/data/logstorage"
-	"link-society.com/flowg/internal/pipelines"
+	"link-society.com/flowg/internal/data/pipelines"
 )
 
 type IngestLogRequest struct {
