+++ api/main.go
@@ -13,7 +13,7 @@ import (
 	"link-society.com/flowg/internal"
 	"link-society.com/flowg/internal/data/auth"
 	"link-society.com/flowg/internal/data/logstorage"
-	"link-society.com/flowg/internal/pipelines"
+	"link-society.com/flowg/internal/data/pipelines"
 )
 
 func NewHandler(
