+++ cmd/flowg/serve.go
@@ -17,7 +17,7 @@ import (
 	"link-society.com/flowg/internal/app/logging"
 	"link-society.com/flowg/internal/data/auth"
 	"link-society.com/flowg/internal/data/logstorage"
-	"link-society.com/flowg/internal/pipelines"
+	"link-society.com/flowg/internal/data/pipelines"
 
 	"link-society.com/flowg/api"
 	"link-society.com/flowg/web"
