+++ web/apps/transformers/controllers/page_new.go
@@ -8,7 +8,7 @@ import (
 	"github.com/a-h/templ"
 
 	"link-society.com/flowg/internal/data/auth"
-	"link-society.com/flowg/internal/pipelines"
+	"link-society.com/flowg/internal/data/pipelines"
 
 	"link-society.com/flowg/web/apps/transformers/templates/views"
 )
